one commit tuesday

+++ src/main/java/com/example/demo/demo/Quiz/repository/QuestionRepository.java
@@ -1,9 +1,8 @@
 package com.example.demo.demo.Quiz.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-
 import com.example.demo.demo.Quiz.model.Question;
 
-public interface QuestionRepository extends JpaRepository<Question, Long>{
-    
+public interface QuestionRepository extends JpaRepository<Question, Integer>{
+    //  Question findByCategory(String category);   
 }
